Fix DeferredFuture crashes by using QWeakPointer
This replaces the custom ref counting with QWeakPointer. It fixes
several issues with the code that lead to crashes.

In the original code there is an inherent race between the
QSharedPointer<DeferredFuture<T>>'s deleter and decWeakRefCount() being
called in the various watch lambdas:

1. If there is even a 1ms pause after line
   https://github.com/benlau/asyncfuture/blob/f2c773bc3d08c43c532d191be5f05617b3c9f088/asyncfuture.h#L712
   it always crashes when there are watchers. This is because the
   strongRefCount is decremented, which means deferred can be deleted
   by a watcher lambda when called. The deleter then crashes on the next
   line as object is no longer valid.

2. If there is a pause before
   https://github.com/benlau/asyncfuture/blob/f2c773bc3d08c43c532d191be5f05617b3c9f088/asyncfuture.h#L689,
   it is possible for the deleteLater to be invoked by the shared
   pointer deleter, and the object deleted in the main thread. This then
   crashes as invokeMethod can't call deleteLater on a deleted object.

3. Calling raw delete this is not thread safe
   (https://github.com/benlau/asyncfuture/blob/f2c773bc3d08c43c532d191be5f05617b3c9f088/asyncfuture.h#L691),
   it should always be doing deleteLater in the presence of multiple
   threads calling deleteLater.

There is not really a good way to fix this with the existing approach.
The best way I found is to use the QSharedPointer semantics to handle
ordered deletion. We can maintain a weakRef internally which can be used
to obtain a shared pointer reference whenever we need to create a new
watcher.

In my testing, this leads to deferred being correctly deleted in all
cases without any race condition.